feat(sql) : adds solution to find the percentage of the users registered in each contest rounded to two decimals.
and return the result table ordered by percentage in descending order. In case of a tie, order it by contest_id in ascending order.
uses an inner join to compare users registered in a contest to all the number of users in the system where the denimenator is a subquery that returns the count of distinct users in the system
orders by percentage in descending order and for ties, by contest_id in ascending order
groups by contest_id to correctly aggregates the data for each contest